renames projects namespace to overview
